Bugfix for nested wrappers in React
An issue occurred with the generated code where if you had to wrappers
using {children} syntax next to eachother like so:
```js
<Wrapper>
  <WrapperTwo>
    Hello
  </WrapperTwo>
<Wrapper>
```
The generated code would add a function block inside of <Wrapper> like
so:
```js
<Wrapper>
  (() => {})
<Wrapper>
```

Which wasn't being properly wrapped in `{}`. The reason for this is that
as the tree changes the parent for the second node <WrapperTwo> wouldn't
be the JSXElement <Wrapper> but instead a "CallExpression".

By reparsing the tree we fix the parent back to the JSXElement.